+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.1.2 / 2019-12-09
+==================
+* [MISC] Updated dev dependencies to their latest versions.
+* [MISC] Removed old console.log statements.
+* [FEATURE] Added getter for the task's function.
+
 0.1.1 / 2019-11-19
 ==================
 * [MISC] Updated all dependencies to their latest versions.